Add portable endpoint facade
