+++ .github/workflows/ci.yml
@@ -58,11 +58,12 @@ jobs:
       - run: cargo clippy -p minip2p-rs --features mdns --all-targets -- -D warnings
       - run: cargo clippy -p minip2p-rs --features discovery,mdns --all-targets -- -D warnings
       - run: cargo clippy -p minip2p-rs --features tcp --all-targets -- -D warnings
+      - run: cargo clippy -p minip2p-rs --no-default-features --features std,tcp --all-targets -- -D warnings
       - run: cargo clippy -p minip2p-rs --features discovery,mdns,tcp --all-targets -- -D warnings
       - run: cargo clippy -p minip2p-tcp --features smoltcp --all-targets -- -D warnings
       - run: cargo clippy -p minip2p-mdns --features smoltcp --all-targets -- -D warnings
       - run: cargo clippy --manifest-path fuzz/Cargo.toml --all-targets -- -D warnings
-      - run: cargo check --no-default-features --target thumbv7em-none-eabi -p minip2p-core -p minip2p-platform -p minip2p-identity -p minip2p-transport -p minip2p-tls -p minip2p-noise -p minip2p-yamux -p minip2p-secure-mux -p minip2p-tcp -p minip2p-circuit -p minip2p-identify -p minip2p-multistream-select -p minip2p-ping -p minip2p-pubsub -p minip2p-discovery -p minip2p-mdns -p minip2p-relay -p minip2p-autonat -p minip2p-dcutr -p minip2p-swarm -p minip2p-nat
+      - run: cargo check --no-default-features --target thumbv7em-none-eabi -p minip2p-core -p minip2p-platform -p minip2p-identity -p minip2p-transport -p minip2p-tls -p minip2p-noise -p minip2p-yamux -p minip2p-secure-mux -p minip2p-tcp -p minip2p-circuit -p minip2p-identify -p minip2p-multistream-select -p minip2p-ping -p minip2p-pubsub -p minip2p-discovery -p minip2p-mdns -p minip2p-relay -p minip2p-autonat -p minip2p-dcutr -p minip2p-swarm -p minip2p-nat -p minip2p-rs
       - run: cargo check --no-default-features --features smoltcp --target thumbv7em-none-eabi -p minip2p-tcp -p minip2p-mdns
       - run: cargo test --workspace
       - run: cargo test -p minip2p-rs --features nat
