+++ crates/minip2p/Cargo.toml
@@ -17,26 +17,37 @@ name = "minip2p"
 workspace = true
 
 [features]
+default = ["std", "quic"]
+std = [
+    "minip2p-core/std",
+    "minip2p-identify/std",
+    "minip2p-identity/std",
+    "minip2p-platform/std",
+    "minip2p-swarm/std",
+    "minip2p-transport/std",
+    "minip2p-tcp?/std",
+]
+quic = ["std", "dep:minip2p-quic"]
 tcp = ["dep:minip2p-tcp"]
-nat = ["dep:minip2p-circuit", "dep:minip2p-nat", "dep:getrandom"]
+nat = ["std", "dep:minip2p-circuit", "dep:minip2p-nat", "dep:getrandom"]
 pubsub = ["dep:minip2p-pubsub", "dep:thiserror"]
 discovery = ["nat", "pubsub", "dep:minip2p-discovery"]
 mdns = ["nat", "dep:minip2p-mdns", "dep:minip2p-discovery", "dep:thiserror"]
 
 [dependencies]
-minip2p-core = { path = "../core", version = "0.3.1" }
+minip2p-core = { path = "../core", version = "0.3.1", default-features = false }
 minip2p-circuit = { path = "../circuit", version = "0.3.1", optional = true }
 minip2p-discovery = { path = "../discovery", version = "0.3.1", optional = true }
-minip2p-identify = { path = "../identify", version = "0.3.1" }
-minip2p-identity = { path = "../identity", version = "0.3.1" }
+minip2p-identify = { path = "../identify", version = "0.3.1", default-features = false }
+minip2p-identity = { path = "../identity", version = "0.3.1", default-features = false, features = ["ed25519"] }
 minip2p-mdns = { path = "../mdns", version = "0.3.1", optional = true }
 minip2p-nat = { path = "../nat", version = "0.3.1", optional = true }
-minip2p-platform = { path = "../platform", version = "0.3.1" }
+minip2p-platform = { path = "../platform", version = "0.3.1", default-features = false }
 minip2p-pubsub = { path = "../pubsub", version = "0.3.1", optional = true }
-minip2p-quic = { path = "../../transports/quic", version = "0.3.1" }
-minip2p-tcp = { path = "../../transports/tcp", version = "0.3.1", optional = true }
-minip2p-swarm = { path = "../swarm", version = "0.3.1" }
-minip2p-transport = { path = "../transport", version = "0.3.1" }
+minip2p-quic = { path = "../../transports/quic", version = "0.3.1", optional = true }
+minip2p-tcp = { path = "../../transports/tcp", version = "0.3.1", default-features = false, optional = true }
+minip2p-swarm = { path = "../swarm", version = "0.3.1", default-features = false }
+minip2p-transport = { path = "../transport", version = "0.3.1", default-features = false }
 getrandom = { version = "0.4.3", optional = true }
 thiserror = { version = "2.0.18", optional = true }
