+++ crates/minip2p/src/discovery.rs
@@ -14,12 +14,13 @@ use minip2p_nat::{ConnectId, NatEvent};
 use minip2p_pubsub::PubsubEvent;
 use minip2p_swarm::SwarmEvent;
 
-use crate::EndpointSwarm;
 #[cfg(feature = "mdns")]
-use crate::mdns::MdnsDriver;
+use super::mdns::MdnsDriver;
+use super::nat::NatDriver;
 #[cfg(feature = "discovery")]
-use crate::pubsub::PubsubDriver;
-use crate::{Error, nat::NatDriver};
+use super::pubsub::PubsubDriver;
+use crate::EndpointSwarm;
+use crate::Error;
 
 /// Errors from discovery-focused endpoint waits.
 #[derive(Debug, thiserror::Error)]
