+++ justfile
@@ -15,6 +15,7 @@ check:
     cargo check -p minip2p-rs --features mdns
     cargo check -p minip2p-rs --features discovery,mdns
     cargo check -p minip2p-rs --features tcp
+    cargo check -p minip2p-rs --no-default-features --features std,tcp
     cargo check -p minip2p-rs --features discovery,mdns,tcp
     cargo check -p minip2p-tcp --features smoltcp --all-targets
     cargo check -p minip2p-mdns --features smoltcp --all-targets
@@ -25,6 +26,7 @@ clippy:
     cargo clippy -p minip2p-rs --features mdns --all-targets -- -D warnings
     cargo clippy -p minip2p-rs --features discovery,mdns --all-targets -- -D warnings
     cargo clippy -p minip2p-rs --features tcp --all-targets -- -D warnings
+    cargo clippy -p minip2p-rs --no-default-features --features std,tcp --all-targets -- -D warnings
     cargo clippy -p minip2p-rs --features discovery,mdns,tcp --all-targets -- -D warnings
     cargo clippy -p minip2p-tcp --features smoltcp --all-targets -- -D warnings
     cargo clippy -p minip2p-mdns --features smoltcp --all-targets -- -D warnings
@@ -45,7 +47,7 @@ test:
 
 check-nostd:
     rustup target add thumbv7em-none-eabi
-    cargo check --no-default-features --target thumbv7em-none-eabi -p minip2p-core -p minip2p-platform -p minip2p-identity -p minip2p-transport -p minip2p-tls -p minip2p-noise -p minip2p-yamux -p minip2p-secure-mux -p minip2p-tcp -p minip2p-circuit -p minip2p-identify -p minip2p-multistream-select -p minip2p-ping -p minip2p-pubsub -p minip2p-discovery -p minip2p-mdns -p minip2p-relay -p minip2p-autonat -p minip2p-dcutr -p minip2p-swarm -p minip2p-nat
+    cargo check --no-default-features --target thumbv7em-none-eabi -p minip2p-core -p minip2p-platform -p minip2p-identity -p minip2p-transport -p minip2p-tls -p minip2p-noise -p minip2p-yamux -p minip2p-secure-mux -p minip2p-tcp -p minip2p-circuit -p minip2p-identify -p minip2p-multistream-select -p minip2p-ping -p minip2p-pubsub -p minip2p-discovery -p minip2p-mdns -p minip2p-relay -p minip2p-autonat -p minip2p-dcutr -p minip2p-swarm -p minip2p-nat -p minip2p-rs
     cargo check --no-default-features --features smoltcp --target thumbv7em-none-eabi -p minip2p-tcp -p minip2p-mdns
 
 peer-ping:
